# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "World UZ",
        "description": "Documentation for World Uz Company",
        "contact": {
            "email": "rustamovramziddin7@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/university": {
            "get": {
                "tags": [
                    "University"
                ],
                "summary": "Return only name and image for the  first page to avoid large date",
                "description": "Return only name and image for the  first page",
                "operationId": "8d0b812cbb321a3715292f228fa0c3bc",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "University"
                ],
                "summary": "Post a new data",
                "description": "Post new University  data",
                "operationId": "5fdf71fc5b125ed93041585e04f04894",
                "requestBody": {
                    "description": "Pass Country   credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image",
                                    "country_id",
                                    "categories",
                                    "contact_file",
                                    "min_price",
                                    "min_ielts",
                                    "city_name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "AQSH"
                                    },
                                    "image": {
                                        "type": "text",
                                        "format": "png,jpg,jpeg,gif",
                                        "example": "image"
                                    },
                                    "country_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    },
                                    "categories": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "Master"
                                    },
                                    "contact_file": {
                                        "type": "file",
                                        "format": "file",
                                        "example": "file.jpeg"
                                    },
                                    "min_price": {
                                        "type": "text",
                                        "format": "string",
                                        "example": "$2000"
                                    },
                                    "min_ielts": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "5.5"
                                    },
                                    "city_name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "New York"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/v1/university/{university}": {
            "get": {
                "tags": [
                    "University"
                ],
                "summary": "Get one with field relation data !",
                "description": "Return all date related to ID{university id}",
                "operationId": "69a3f61f1ae67a8dcea68a95a785c410",
                "parameters": [
                    {
                        "name": "university",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "University"
                ],
                "summary": "Get one and Delete related to university id",
                "description": "Return  date related to ID of the university",
                "operationId": "d1260271fb71e10069dc3282281342b2",
                "parameters": [
                    {
                        "name": "university",
                        "in": "path",
                        "description": "put university id and try to delete ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/v1/universities/{university}/field": {
            "get": {
                "tags": [
                    "University"
                ],
                "summary": "Get one return with relation data from database like fields all the field come here ! ",
                "description": "Return all date related to ID{university id}",
                "operationId": "951be325b620bba43ac1a4b2d34f8a1d",
                "parameters": [
                    {
                        "name": "university",
                        "in": "path",
                        "description": "University id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}